Add files via upload
Softly cycle your window colors in Windows 7! Let everyone around you know how absolutely fabulous you are! 

GAY_AF is a script that hijacks the DWM API and pushes window color changes. 
It contains an HSL to RGB to DWM-API converter, and in theory can input and output any HSL or RGB code, I'm considering tying it to my Philips Hue system. 

I've also written in a mouse tracker that makes the hue cycling faster during mouse movement, making it seem like the computer is almost excited that you're using it. This can be disabled in the constant values section.